src/main/java/com/example/EmployeePlural.java
package com.example;

import java.util.List;

public class EmployeePlural {

	private final EmployeeRepository employeeRepository;
	private final BankService bankService;

	public EmployeePlural(EmployeeRepository employeeRepository, BankService bankService) {
		this.employeeRepository = employeeRepository;
		this.bankService = bankService;
	}

	public int payEmployees() {
		List<Employee> employees = employeeRepository.findAll();
		int payments = 0;
		for (Employee employee : employees) {
			try {
				bankService.pay(employee.getId(), employee.getSalary());
				employee.setPaid(true);
				payments++;
			} catch (RuntimeException e) {
				employee.setPaid(false);
			}
		}
		return payments;
	}
}
